// apps/crm/client/src/features/scrumboard/components/ScrumboardColumnOptionsBtn.module.scss.d.ts
export type Styles = {
  scrumboardOptionBtn: string;
  scrumboardOptionBtn__menu: string;
  scrumboardOptionBtn__menuItem: string;
  scrumboardOptionBtn__menuItem__svg: string;
  scrumboardOptionBtn__menuItemWarning: string;
  scrumboardOptionBtn__menuItemWarning__svg: string;
  scrumboardOptionBtn__popover: string;
  scrumboardOptionBtn__svg: string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;